application/pyproject.toml
[tool.poetry]
name = "fast_api_practice"
version = "1.0.1"
description = "テンプレート"
authors = ["takuron1996"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { version = "^0.103.1", extras = ["all"] }
SQLAlchemy = "^2.0.21"
psycopg2 = "^2.9.9"
asyncpg = "^0.28.0"
alembic = "^1.12.0"
ulid-py = "^1.1.0"
bcrypt = "^4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-clarity = "^1.0.1"
pytest-randomly = "^3.15.0"
pytest-mock = "^3.11.1"
pytest-xdist = "^3.3.1"
pytest-bdd = "^6.1.1"
pytest-env = "^1.0.1"
ruff = "^0.0.291"
pdoc = "^14.1.0"
mypy = "^1.5.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# .coveragercの設定
[tool.coverage.run]
branch = true
source = ["practice/", "config/"]
omit = ["*/__init__.py", "*/tests/*"]

[tool.coverage.report]
# Regexes for lines to exclude from consideration
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if self\\.debug",
  "raise AssertionError",
  "raise NotImplementedError",
  "if 0:",
  "if __name__ == .__main__.:",
  "@(abc\\.)?abstractmethod",
]
ignore_errors = true


[tool.coverage.html]
directory = "htmlcov"

[tool.mypy]
packages = ["practice", "config"]
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E", "F", "W", "D", "I"]
ignore = []

extend-exclude = ["migrations/versions/"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]

line-length = 80

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.11
target-version = "py311"

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.per-file-ignores]
"__init__.py" = ["ALL"]
